refactor: single asteroid server component to server-client-server component
