+++ next.config.js
@@ -1,2 +1,6 @@
 /** @type {import('next').NextConfig} */
-module.exports = {};
+module.exports = {
+  experimental: {
+    serverActions: true,
+  },
+};
